+++ services/video/main.c
@@ -108,7 +108,7 @@ s32 main(void) {
 								s8 *buf = (s8*)malloc(msg.length * sizeof(s8));
 								read(fd,buf,msg.length);
 								startPos = (u16*)buf;
-								if(msg.length - sizeof(u16) <= *startPos * 2 + COLS * ROWS * 2)
+								if(msg.length - sizeof(u16) <= (u32)*startPos * 2 + COLS * ROWS * 2)
 									vid_setScreen(*startPos,buf + sizeof(u16),msg.length - sizeof(u16));
 								free(buf);
 							}
